backend/foodgram/requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-filter==2.4.0
gunicorn==20.0.4
sqlparse==0.3.1
django-colorfield==0.5.0
Pillow==8.4.0
djoser
django-filter
xhtml2pdf